os path to db, add files .pkl .db .xlsx
